# Scan memory reuse