--- a/theano/scan_module/scan_op.py
+++ b/theano/scan_module/scan_op.py
@@ -675,7 +675,7 @@ class Scan(PureOp):
                   self.n_sit_sot +
                   self.n_nit_sot)
         wrapped_inputs = [Param(x, borrow=False) for x in self.inputs]
-        wrapped_outputs = [Out(x, borrow=(x not in self.inputs)) for x in
+        wrapped_outputs = [Out(x, borrow=True) for x in
                            self.outputs[:slices]]
         wrapped_outputs += self.outputs[slices:]
         profile = None