… theano/scan_module/scan_perform.c