This PR fixes two bug in the currently deactivated memory pre-allocation feature in Scan and activates it in both the Python and Cython backends. This means that is doesn't fix bugs that are currently occurring but rather it fixes the memory pre-allocation so that we can use it without introducing bugs.

The first bug it fixes comes from an optimization in `GpuDimshuffle` in the cuda gpu backend where, to avoid creating a new array, it can reuse the one it receives as input and alter its attributes, including the `gpudata` attribute. This caused an issue in the old code for the memory pre-allocation because it relied on the assumption that a given CudaNdarray object would always point to the same memory region and that if an Op needed to return a CudaNdarray pointing to a new memory region, it would create a new one and leave the original untouched.

The second bug comes from the SaveMem optimization and the fact that it tries (if it can) to keep a circular buffer only as small as the number of states required at any given iteration to compute the next state. This works without the memory reuse feature because all the outputs of the inner function are computed and then they are put in the buffers. With the memory reuse, since the outputs are stored in the buffers as they are computed by the inner function, the buffers need to be at least 1 element bigger so that the new states can be stored without overwriting the oldest state that is still used in the inner function.

Now, there are added checks to ensure that the pre-allocated array has been reused *and* that it still points to the same data.