@carriepl 2015-05-07
I benchmarked the latest commit on the Deeplearning.net LSTM tutorial on a GTX470. 

Under the Cython backend, time per call for the fprop Scan drops from 0.026s to 0.013s (overhead % drops from 45% to 5%). For the bprop scan, time per call drops from 0.117s to 0.093s (overhead % drops from 25% to 13%). Globally, the time taken to train for one epoch drop from 11.26s to 7.63s.