@nouiz 2015-05-07
I think it is good, but there is a part I can't review. Can you document all that at one place? Otherwise, we will all forget the details when we need to play with this next time.