@carriepl 2015-05-08
I added the documentation in the scan_op.py file.

@abergeron and/or @lamblin, can one of you take a look? The part in ScanSaveMem probably deserves even more attention than the others since Fred didn't review it.