@nouiz 2015-05-08
What about making a Theano flag scan.prealloc_output that would disable this to not cause extra memory usage?
I discussed yesterday with @carriepl about adding this as an option in the info dict of scan, and make an opt that swap enable it to make DebugMode able to test this. But he wasn't sure if this would end-up as what happened with the gpua key in that dict. @abergeron what do you think?