@nouiz 2015-05-08
Also, I asked a 3 persons to test it with big model to see if the memory increase cause problems. If so, it make it more important to allow to disable this.