@abergeron 2015-05-08
I don't see why DebugMode should be able to mess with that.  It's a very op-specific setting.

Also, if we add a dict option it's very easy to forget to transfer it at one spot and have bugs that way.