@sebastien-j 2015-05-10
I tested this pull request for the largest NMT models I trained on a 3GB GPU (approximately 15k source and target words). The memory usage is fine. For the largest minibatches, it now takes ~2.03 seconds instead of ~2.095s.