@carriepl 2015-05-11
Modified the test to use float32