@carriepl 2015-05-11
Added new flag (scan.allow_output_prealloc) to allow/disallow new optimization