@abergeron 2015-05-12
This looks fine but another set of eyes wouldn't be too much.