@carriepl 2015-05-12
Clarified comment for section 4.5 in both Scan backends and fixed the documentation for the flag scan.allow_output_prealloc.